user login and signup

+++ src/main/java/controller/DeleteMovie.java
@@ -9,6 +9,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import dao.Dao;
 @WebServlet("/deletemovie")
@@ -19,10 +20,21 @@ protected void service(HttpServletRequest req, HttpServletResponse resp) throws 
 		int id = Integer.parseInt(req.getParameter("id"));
 		Dao dao = new Dao();
 		try {
+			HttpSession session =req.getSession();
+			String adminname=(String )session.getAttribute("adminname");
+			if(adminname!=null) {
+				
 			dao.deleteMovie(id);
 			req.setAttribute("movies", dao.getAllMovie());
 			RequestDispatcher dispatcher = req.getRequestDispatcher("home.jsp");
 			dispatcher.include(req, resp);
+			}
+			else {
+				req.setAttribute("msgp", "access denied, admin login is reqired" );
+				RequestDispatcher dispatcher = req.getRequestDispatcher("alogin.jsp");
+				dispatcher.include(req, resp);
+				
+			}
 			
 			
 		} catch (ClassNotFoundException e) {
